move constantes to common pkg
