--- api/common/dbconsts/consts.go
@@ -1,7 +1,0 @@
-package dbconsts
-
-// influx-db constants
-const (
-	ContributorsActivityMeasurement = "contributors_activity"
-	ContributorsStatsMeasurement    = "contributors_stats"
-)
