+++ common/dbconsts/consts.go
@@ -1,1 +1,7 @@
 package dbconsts
+
+// influx-db constants
+const (
+	ContributorsActivityMeasurement = "contributors_activity"
+	ContributorsStatsMeasurement    = "contributors_stats"
+)
