+++ jobs/jobs/stats/stats.go
@@ -8,7 +8,7 @@ import (
 	"github.com/influxdata/influxdb-client-go/v2/api"
 	"github.com/influxdata/influxdb-client-go/v2/api/write"
 	"github.com/pkg/errors"
-	"github.com/wormhole-foundation/wormhole-explorer/api/common/dbconsts"
+	"github.com/wormhole-foundation/wormhole-explorer/common/dbconsts"
 	"go.uber.org/zap"
 	"io"
 	"net/http"
